requirements.txt
transformers
tensorflow-cpu
numpy
seaborn
matplotlib
sklearn